requirements.txt
Flask
Pillow
gunicorn
pymongo[srv]
